+++ CHANGELOG.rst
@@ -2,6 +2,7 @@
     Features:
 
     * ``SignalReader`` base class and ``CsvReader``.
+    * ``Signal.plot``: grid on by default. New parameter ``t0``.
 
     Fixes: